Refactor settings actions, reducers, and state.
