+++ src/settings/settingsState.js
@@ -1,0 +1,5 @@
+const SettingsState = {
+  settingsUoM: 'oz' // unit of measurement
+};
+
+export default SettingsState;
