+++ src/state/reducers.js
@@ -1,33 +1,9 @@
 import { combineReducers } from 'redux';
 
-import {
-  CHANGE_WEED_UOM,
-} from './actions';
-
 import garden from '../garden/gardenReducers';
 import notifications from '../notifications/notificationsReducers';
 import player from '../player/playerReducers';
-
-const initialSettings = {
-  settingsUoM: 'oz' // unit of measurement
-};
-
-/**
- * Settings reducers
- * @param settings
- * @param action
- */
-function settings(settings = initialSettings, action = null) {
-  switch (action.type) {
-    case CHANGE_WEED_UOM:
-      return {
-        ...settings,
-        settingsUoM: action.uom
-      };
-    default:
-      return settings;
-  }
-}
+import settings from '../settings/settingsReducers';
 
 export default combineReducers({
   garden,
